+++ src/main/java/net/loadbang/osc/PacketConsumer.java
@@ -3,10 +3,11 @@
 
 package net.loadbang.osc;
 
+import java.net.InetSocketAddress;
 import net.loadbang.osc.exn.DataException;
 
 /**	A consumer of packets of bytes. */
 
 public interface PacketConsumer {
-	void consumePacket(byte[] packet) throws DataException;
+	void consumePacket(InetSocketAddress source, byte[] packet) throws DataException;
 }